refactor: extract reservation logic into custom hook useReservationManager
